Move the platform path code into a `platform` module
